requirements.txt
google-genai>=0.8.0
requests>=2.25.0
pandas>=1.3.0
